captives & cargo cba settings
