+++ addons/captives/config.cpp
@@ -12,7 +12,7 @@ class CfgPatches {
     };
 };
 
-#include "ACE_Settings.hpp"
+#include "CBA_Settings.hpp"
 #include "CfgEventHandlers.hpp"
 #include "CfgMoves.hpp"
 #include "CfgVehicles.hpp"
