--- addons/cargo/ACE_Settings.hpp
@@ -1,9 +1,0 @@
-class ACE_Settings {
-    class GVAR(enable) {
-        displayName = CSTRING(ModuleSettings_enable);
-        description = CSTRING(ModuleSettings_enable_Description);
-        typeName = "BOOL";
-        value = 1;
-        category = ECSTRING(OptionsMenu,CategoryLogistics);
-    };
-};
