+++ addons/cargo/config.cpp
@@ -12,7 +12,7 @@ class CfgPatches {
     };
 };
 
-#include "ACE_Settings.hpp"
+#include "CBA_Settings.hpp"
 #include "CfgEventHandlers.hpp"
 #include "CfgVehicles.hpp"
 #include "menu.hpp"
